Average for temp/humidity

+++ CircularBuffer/CircularBuffer.h
@@ -7,11 +7,11 @@
 class CircularBuffer {
 public:
     CircularBuffer();
-    void add(uint16_t value);
-    uint16_t getAverage();
+    void add(float value);
+    float getAverage();
 
 private:
-    uint16_t _buffer[MAX_LENGTH];
+    float _buffer[MAX_LENGTH];
     uint8_t _position = 0;
     uint8_t _length = 0;
 };
